split algorithm/detail/search_algorithm_util into common, bfs_impl, dfs_impl
